Fix: Refactor codes for Signup.js file To improve the structure and apply Styled-components appropriately Ref: #134
